uniblock_backend/requirements.txt
anyio==4.6.0
asgiref==3.8.1
base58==2.1.1
certifi==2024.8.30
Deprecated==1.2.14
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
ECPy==1.2.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
inflection==0.5.1
nest-asyncio==1.6.0
packaging==24.1
pycryptodome==3.20.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sniffio==1.3.1
sqlparse==0.5.1
types-Deprecated==1.2.9.20240311
typing_extensions==4.12.2
uritemplate==4.1.1
websockets==11.0.3
wrapt==1.16.0
xrpl-py==3.0.0
